Remove legacy stats
